requirements.txt
annotated-types==0.6.0
boto3==1.34.108
botocore==1.34.108
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
deeplake==3.9.7
dill==0.3.8
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
humbug==0.3.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
lz4==4.3.3
map-boxes==1.0.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pathos==0.3.2
pillow==10.2.0
pip==23.2.1
pox==0.3.4
ppft==1.7.6.8
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
scipy==1.13.0
seaborn==0.13.2
setuptools==68.2.0
six==1.16.0
sympy==1.12
tbb==2021.12.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.18
urllib3==2.2.1
wheel==0.41.2
